requirements/requirements-essential.txt
# Django==1.11.21
setuptools
certifi
django-mailman == 0.4+eso1
mailchimp3 == 3.0.6

# PostgreSQL python client
psycopg2-binary==2.7.7

# Asynchronous Task Queue
celery==4.3.0

# Djangoplicity actions
git+https://@github.com/djangoplicity/djangoplicity-actions@develop

# Djangoplicity
git+https://@github.com/djangoplicity/djangoplicity.git@develop
